### 2018-05-29 510e4d7
add arguments that are passed to `distance()` and update docu
