+++ README.md
@@ -21,7 +21,7 @@ the people who share it with me.
 
 ### Installation
 ```r
-# install philentropy version 0.1.0 from CRAN
+# install philentropy version 0.2.0 from CRAN
 install.packages("philentropy")
 ```
 
