### 2018-06-06 #765 @whedon
[REVIEW]: Philentropy: Information Theory and Distance Quantification with R
**Submitting author:** @HajkD (<a href="http://orcid.org/0000-0002-1567-306X">Hajk-Georg Drost</a>)
**Repository:** <a href="https://github.com/HajkD/philentropy" target ="_blank">https://github.com/HajkD/philentropy</a>
**Version:** v0.2.0
