### 2018-06-06 #752 @leeper
@kellieotto Thanks for agreeing to review for JOSS. The review ticket is open here: https://github.com/openjournals/joss-reviews/issues/765
