### 2018-06-06 #752 @arfon
> @arfon it appears @whedon has become unresponsive.

Looks like he was a little sleepy there for a while...
