### 2018-06-06 8312fa3
fix journal reference according to issue #6
