+++ paper.bib
@@ -1,10 +1,10 @@
 @article{Cha2007,
   title={Comprehensive survey on distance/similarity measures between probability density functions},
   author={Cha, Sung-Hyuk},
-  journal={City},
+  journal={International Journal of Mathematical Models and Methods in Applied Science},
   volume={1},
-  number={2},
-  pages={1},
+  number={4},
+  pages={300-307},
   year={2007}
 }
 
