### 2018-06-07 HajkD/philentropy#8 @kellieotto
Missing LICENSE file
This is detailed feedback for your JOSS review openjournals/joss-reviews#765.

I see that your code and DESCRIPTION contain GPL license info. This is great. JOSS requires you to have a plain-text LICENSE file, in your main repository, with the contents of your chosen software license.
