### 2018-06-07 HajkD/philentropy#9 @kellieotto
Performance claims in the distance vignette
This is detailed feedback for your JOSS review openjournals/joss-reviews#765.

I was able to replicate everything except the last bit using `microbenchmark` to time the different functions. The problem is with running `euclidean`. If I run the code as written I get the error

```
Error in microbenchmark(distance(x, method = "euclidean", test.na = FALSE),  :
  could not find function "euclidean"
```

If I add `philentropy::` before the function call, I get a different error.

```
Error: 'euclidean' is not an exported object from 'namespace:philentropy'
```
