### 2018-06-07 #765 @kellieotto
I've completed my first review. I noted issues with the license and performance claims in issues in the repository.
