### 2018-06-07 #765 @HajkD
Dear Kellie,

Thank you so so much for compiling a review so quickly.
It is very much appreciated.

I will address all your points in detail and will add the missing `LICENSE` file.

Many thanks and kind regards,
Hajk
