### 2018-06-07 161b14c
adding GPL-2 license file according to JOSS policy (see issue #8)
