### 2018-06-07 HajkD/philentropy#8 @HajkD
Dear Kellie,

Many thanks for making me aware of the missing license.
I now added a GLP-2 license file according to JOSS policy.

Kind regards,
Hajk
