### 2018-06-07 a86ed8d
internal (Rcpp) distance metrics were not exported to the philentropy namespace due to a missing `'` in the comments that caused roxygen to omit export of these functions (see issue #9)
