+++ src/distances.h
@@ -49,7 +49,7 @@ double custom_log10(const double& x ){
   }
 }
 
-// @export
+//' @export
 // [[Rcpp::export]]
 double euclidean(const Rcpp::NumericVector& P,const Rcpp::NumericVector& Q, bool testNA){
         
@@ -79,7 +79,7 @@ double euclidean(const Rcpp::NumericVector& P,const Rcpp::NumericVector& Q, bool
         return sqrt(dist);
 }
 
-// @export
+//' @export
 // [[Rcpp::export]]
 double manhattan(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, bool testNA){
         
@@ -111,7 +111,7 @@ double manhattan(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, boo
 
 
 
-// @export
+//' @export
 // [[Rcpp::export]]
 double minkowski(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q,double n, bool testNA){
         
@@ -142,7 +142,7 @@ double minkowski(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q,doub
 }
 
 
-// @export
+//' @export
 // [[Rcpp::export]]
 double chebyshev(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, bool testNA){
         
@@ -175,7 +175,7 @@ double chebyshev(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, boo
 }
 
 
-// @export
+//' @export
 // [[Rcpp::export]]
 double sorensen(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, bool testNA){
         
@@ -218,7 +218,7 @@ double sorensen(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, bool
 
 
 
-// @export
+//' @export
 // [[Rcpp::export]]
 double gower(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, bool testNA){
         
@@ -254,7 +254,7 @@ double gower(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, bool te
 
 
 
-// @export
+//' @export
 // [[Rcpp::export]]
 double soergel(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, bool testNA){
         
@@ -304,7 +304,7 @@ double soergel(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, bool 
 }
 
 
-// @export
+//' @export
 // [[Rcpp::export]]
 double kulczynski_d(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, bool testNA){
         
@@ -362,7 +362,7 @@ double kulczynski_d(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, 
 }
 
 
-// @export
+//' @export
 // [[Rcpp::export]]
 double canberra(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, bool testNA){
         
@@ -407,7 +407,7 @@ double canberra(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, bool
 }
 
 
-// @export
+//' @export
 // [[Rcpp::export]]
 double lorentzian(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, bool testNA, const Rcpp::String unit){
         
@@ -459,7 +459,7 @@ double lorentzian(const Rcpp::NumericVector& P, const Rcpp::NumericVector& Q, bo
 }
 
 
-// @export
+//' @export
 // [[Rcpp::export]]
[…]
