### 2018-06-07 4c78dac
add internal (Rcpp) metrics to philentropy namespace via roxygen (see issue #9)
