+++ NAMESPACE
@@ -6,12 +6,56 @@ export(JE)
 export(JSD)
 export(KL)
 export(MI)
+export(additive_symm_chi_sq)
+export(avg)
+export(bhattacharyya)
 export(binned.kernel.est)
+export(canberra)
+export(chebyshev)
+export(clark_sq)
+export(cosine_dist)
+export(czekanowski)
+export(dice_dist)
 export(dist.diversity)
 export(distance)
+export(divergence_sq)
 export(estimate.probability)
+export(euclidean)
+export(fidelity)
 export(gJSD)
 export(getDistMethods)
+export(gower)
+export(harmonic_mean_dist)
+export(hellinger)
+export(inner_product)
+export(intersection_dist)
+export(jaccard)
+export(jeffreys)
+export(jensen_difference)
+export(jensen_shannon)
+export(k_divergence)
+export(kulczynski_d)
+export(kullback_leibler_distance)
+export(kumar_hassebrook)
+export(kumar_johnson)
 export(lin.cor)
+export(lorentzian)
+export(manhattan)
+export(matusita)
+export(minkowski)
+export(motyka)
+export(neyman_chi_sq)
+export(pearson_chi_sq)
+export(prob_symm_chi_sq)
+export(ruzicka)
+export(soergel)
+export(sorensen)
+export(squared_chi_sq)
+export(squared_chord)
+export(squared_euclidean)
+export(taneja)
+export(tanimoto)
+export(topsoe)
+export(wave_hedges)
 importFrom(Rcpp,sourceCpp)
 useDynLib(philentropy, .registration = TRUE)
