### 2018-06-07 c5b91cb
add exported Rcpp code (see issue #9)
