+++ R/RcppExports.R
@@ -1,6 +1,22 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+Ecpp <- function(P, unit) {
+    .Call(`_philentropy_Ecpp`, P, unit)
+}
+
+JEcpp <- function(JointProbabilities, unit) {
+    .Call(`_philentropy_JEcpp`, JointProbabilities, unit)
+}
+
+CEcpp <- function(JointProbabilities, Probabilities, unit) {
+    .Call(`_philentropy_CEcpp`, JointProbabilities, Probabilities, unit)
+}
+
+MIcpp <- function(X, Y, XY, unit) {
+    .Call(`_philentropy_MIcpp`, X, Y, XY, unit)
+}
+
 pearson_corr_centred <- function(x, y, testNA) {
     .Call(`_philentropy_pearson_corr_centred`, x, y, testNA)
 }
@@ -41,196 +57,224 @@ custom_log10 <- function(x) {
     .Call(`_philentropy_custom_log10`, x)
 }
 
+#' @export
 euclidean <- function(P, Q, testNA) {
     .Call(`_philentropy_euclidean`, P, Q, testNA)
 }
 
+#' @export
 manhattan <- function(P, Q, testNA) {
     .Call(`_philentropy_manhattan`, P, Q, testNA)
 }
 
+#' @export
 minkowski <- function(P, Q, n, testNA) {
     .Call(`_philentropy_minkowski`, P, Q, n, testNA)
 }
 
+#' @export
 chebyshev <- function(P, Q, testNA) {
     .Call(`_philentropy_chebyshev`, P, Q, testNA)
 }
 
+#' @export
 sorensen <- function(P, Q, testNA) {
     .Call(`_philentropy_sorensen`, P, Q, testNA)
 }
 
+#' @export
 gower <- function(P, Q, testNA) {
     .Call(`_philentropy_gower`, P, Q, testNA)
 }
 
+#' @export
 soergel <- function(P, Q, testNA) {
     .Call(`_philentropy_soergel`, P, Q, testNA)
 }
 
+#' @export
 kulczynski_d <- function(P, Q, testNA) {
     .Call(`_philentropy_kulczynski_d`, P, Q, testNA)
 }
 
+#' @export
 canberra <- function(P, Q, testNA) {
     .Call(`_philentropy_canberra`, P, Q, testNA)
 }
 
+#' @export
 lorentzian <- function(P, Q, testNA, unit) {
     .Call(`_philentropy_lorentzian`, P, Q, testNA, unit)
 }
 
+#' @export
 intersection_dist <- function(P, Q, testNA) {
     .Call(`_philentropy_intersection_dist`, P, Q, testNA)
 }
 
+#' @export
 wave_hedges <- function(P, Q, testNA) {
     .Call(`_philentropy_wave_hedges`, P, Q, testNA)
 }
 
+#' @export
 czekanowski <- function(P, Q, testNA) {
     .Call(`_philentropy_czekanowski`, P, Q, testNA)
 }
 
+#' @export
 motyka <- function(P, Q, testNA) {
     .Call(`_philentropy_motyka`, P, Q, testNA)
 }
 
+#' @export
 tanimoto <- function(P, Q, testNA) {
     .Call(`_philentropy_tanimoto`, P, Q, testNA)
 }
 
+#' @export
 ruzicka <- function(P, Q, testNA) {
     .Call(`_philentropy_ruzicka`, P, Q, testNA)
 }
 
+#' @export
 inner_product <- function(P, Q, testNA) {
[…]
