### 2018-06-07 HajkD/philentropy#9 @HajkD
Dear Kellie,

I am very sorry for the confusion.  There were missing `'` in my comments (see file `src/distances.h`) so that the `roxygen` package did not automatically export my functions to the `philentropy` namespace.
Since in my developer environment, these functions were defined I could use them, so I didn't
see this issue.

Thus, all I had to do was to exchange the comments:

```r
// @export
// [[Rcpp::export]]
```

by

```r
//' @export
// [[Rcpp::export]]
```

This way `roxygen` exported all internal metric functions to the `philentropy` namespace.
The benchmark code and example should be working now.

Please accept my apologies for the dysfunctional example.

Many thanks and kind regards,
Hajk
