### 2018-06-07 a540546
update travis options
