+++ .travis.yml
@@ -1,6 +1,6 @@
 language: r
 bioc_required: true
-warnings_are_errors: true
+warnings_are_errors: false
 sudo: false
 cache: packages
 r_check_revdep: false
