### 2018-06-08 #765 @HajkD
Dear all,

I incorporated all changes and addressed the two issues listed above.

Many thanks and kind regards,
Hajk
