### 2018-06-09 HajkD/philentropy#9 @kellieotto
It works now when I install the version on GitHub. Just make sure to update the CRAN version eventually!
