### 2018-06-09 #765 @leeper
@HajkD Please issue a GitHub release for the current version of the software, and create a DOI for that release using, for example, figshare or Zenodo. Once you have the DOI, please leave it here as a comment. Once that's done, your paper will be accepted at JOSS!

@kellieotto Thanks so much for your very quick review!
