### 2018-06-09 0856354
update to new version for JOSS doi
