### 2018-06-09 #765 @HajkD
Dear @leeper,

Thank you so much for organizing everything.

Here is the doi I generated:

https://doi.org/10.5281/zenodo.1286221

@kellieotto Thank you very very much for your great and super fast review.

This was a great experience and I will definitely recommend JOSS!

Kind regards,
Hajk
