### 2018-06-10 0319aa6
update NEWS
