### 2018-06-11 #765 @HajkD
Dear all,

Thank you so so much for your great work and for the super fast editorial/review process.

This is my first JOSS paper and I will definitely consider submitting future tools to JOSS as well. In addition, I will also highly recommend it to my colleagues.

Kind regards,
Hajk
